ZPI adjust to verify happy path  - rename onlaunch package  - add organization prefix to redirections  - adjust createclientanduser script
