+++ scripts/createclientanduser.http
@@ -43,7 +43,7 @@ Content-Type: application/json
 
 ###
 
-POST http://localhost:8080/api/user/register
+POST http://localhost:8080/api/organization/pizza-house/user/register
 accept: */*
 Content-Type: application/json
