+++ src/main/java/com/zpi/api/ui/UIRedirect.java
@@ -6,10 +6,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 public class UIRedirect {
 
-    public static final String ALLOW_URI = "/allow/**";
-    public static final String SIGN_IN_URI = "/signin/**";
-    public static final String SIGN_UP_URI = "/signup/**";
-    public static final String DASHBOARD = "/dashboard/**";
+    private static final String ORGANIZATION_URI = "/organization/*";
+    public static final String ALLOW_URI =  "/allow/**";
+    public static final String SIGN_IN_URI =  ORGANIZATION_URI + "/signin/**";
+    public static final String SIGN_UP_URI =  ORGANIZATION_URI + "/signup/**";
+    public static final String DASHBOARD =  ORGANIZATION_URI + "/dashboard/**";
 
     private static final String INDEX_URI = "forward:/index.html";
