+++ src/test/groovy/com/zpi/CommonFixtures.groovy
@@ -23,7 +23,7 @@ class CommonFixtures {
     public static final String password = "Password"
 
     public static final String ticket = "defaultTicketsfasdgfartasdfafta"
-    public static final String authPageUrl = "/signin"
+    public static final String authPageUrl = "/organization/signin"
     public static final String grantType = "authorization_code"
     public static final String hardcodedCode = "asdf"
